+++ lib/src/widgets/containers/lit_settings_footer.dart
@@ -13,16 +13,21 @@ class LitSettingsFooter extends StatelessWidget {
   /// A list of buttons to create interactive options.
   final List<LitPlainLabelButton> children;
 
+  /// A list of colors applied on the background's gradient.
+  final List<Color> gradientColors;
+
   /// Creates a [LitSettingsFooter].
   const LitSettingsFooter({
     Key? key,
     this.title = "Settings",
     required this.children,
+    this.gradientColors = LitFooterDefaultStyling.colorsLight,
   }) : super(key: key);
 
   Widget build(BuildContext context) {
     return LitFooter(
       title: title,
+      gradientColors: gradientColors,
       child: Container(
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.end,